Fix our epoch
This turned into a bit of an epic, but it implements the changes we
discussed about site-state clearing, epoch start time, and all the
associated logic.

This defines when epochs start (randomly, as agreed).

This defines a store for that per-site state and integrates a new
algorithm for getting the current epoch.

It defines a persistent clear-state time variable that tracks when any
site state was cleared.  This value blocks out use of the API globally
for a period.

I've added notes about partitioning the state-cleared variable, but not
specified anything for that.  That seems like a reasonable balance to
me; as the text says, this is the most private option.

Closes #17.