Remove unused stuff
This includes dropping the somewhat non-specific discussion about epochs
from the "clearing site data" part.  It is enough to say "all" at this
point.  User agents are able to make their own decisions about what to
do with data that is prior to the earliest epoch (and nothing can be
said about future epochs).