Add text about the choice of clock
And an analysis of why our choice is acceptable.

I didn't have to rely on the fact that most clock adjustments are small,
except somewhat indirectly.  Instead, it mentions that large forward
adjustments that might cause a change of epoch each have an
instantaneous effect on privacy by refreshing the budget.