Deal with privacy budget rounding
This converts the underlying type from a CSS float (yes, we were citing
CSS, not a numeric type) to a 32-bit integer.  I figure that an epsilon
of 4000 is unnecessary, so this seemed like a nice way to manage it.

The algorithm is slightly different and I've tweaked some of the
references to it.  Most of that is just referencing and cleanup.

Closes #77.