Spec the multi-epoch unoptimization
Prior to this, the algorithm for deducting privacy budget used the L1
norm of the histogram, which the analysis shows is only possible when
impressions from a single epoch are involved.

Again, this is pretty suboptimal if implemented directly, which is a
recurring pattern.  This iterates over all impressions twice.  I didn't
bother to cache the impressions that were selected.  I didn't even
bother to break out of the loop when the count hits two.  Those would
just make the "code" harder to read than it already is.

This is contrary to what is implied in #78.
Still, this closes #78, just in the opposite way to what was invisaged.