Encryption for DAP
This does a more than thorough definition of what it takes to encrypt a
report for DAP/MPC.  This is probably a little TOO detailed, in that the
DAP draft specifies most of this.  However, I found it helpful to go to
this level of detail because the draft doesn't pin down certain key
parameters, like the choice of field and time resolution.